much better that this lives back wih logging
