+++ logging/logrus/ctxlogrus/examples_test.go
@@ -1,24 +1,22 @@
-package ctx_zap_test
+package ctxlogrus_test
 
 import (
-	"context"
-
+	"github.com/grpc-ecosystem/go-grpc-middleware/logging/logrus/ctxlogrus"
 	"github.com/grpc-ecosystem/go-grpc-middleware/tags"
-	"github.com/grpc-ecosystem/go-grpc-middleware/tags/zap"
 	pb_testproto "github.com/grpc-ecosystem/go-grpc-middleware/testing/testproto"
-	"go.uber.org/zap"
+	"github.com/sirupsen/logrus"
+	"golang.org/x/net/context"
 )
 
-var zapLogger *zap.Logger
+var logrusLogger *logrus.Logger
 
 // Simple unary handler that adds custom fields to the requests's context. These will be used for all log statements.
 func Example_HandlerUsageUnaryPing() {
 	_ = func(ctx context.Context, ping *pb_testproto.PingRequest) (*pb_testproto.PingResponse, error) {
 		// Add fields the ctxtags of the request which will be added to all extracted loggers.
 		grpc_ctxtags.Extract(ctx).Set("custom_tags.string", "something").Set("custom_tags.int", 1337)
-
-		// Extract a single request-scoped zap.Logger and log messages.
-		l := ctx_zap.Extract(ctx)
+		// Extract a single request-scoped logrus.Logger and log messages.
+		l := ctxlogrus.Extract(ctx)
 		l.Info("some ping")
 		l.Info("another ping")
 		return &pb_testproto.PingResponse{Value: ping.Value}, nil
