+++ logging/zap/context.go
@@ -1,7 +1,7 @@
 package grpc_zap
 
 import (
-	"github.com/grpc-ecosystem/go-grpc-middleware/tags/zap"
+	"github.com/grpc-ecosystem/go-grpc-middleware/logging/zap/ctxzap"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 	"golang.org/x/net/context"
@@ -10,11 +10,11 @@ import (
 // AddFields adds zap fields to the logger.
 // Deprecated: should use the ctx_zap.AddFields instead
 func AddFields(ctx context.Context, fields ...zapcore.Field) {
-	ctx_zap.AddFields(ctx, fields...)
+	ctxzap.AddFields(ctx, fields...)
 }
 
 // Extract takes the call-scoped Logger from grpc_zap middleware.
 // Deprecated: should use the ctx_zap.Extract instead
 func Extract(ctx context.Context) *zap.Logger {
-	return ctx_zap.Extract(ctx)
+	return ctxzap.Extract(ctx)
 }
