+++ logging/zap/ctxzap/examples_test.go
@@ -1,22 +1,24 @@
-package ctx_logrus_test
+package ctxzap_test
 
 import (
+	"context"
+
+	"github.com/grpc-ecosystem/go-grpc-middleware/logging/zap/ctxzap"
 	"github.com/grpc-ecosystem/go-grpc-middleware/tags"
-	"github.com/grpc-ecosystem/go-grpc-middleware/tags/logrus"
 	pb_testproto "github.com/grpc-ecosystem/go-grpc-middleware/testing/testproto"
-	"github.com/sirupsen/logrus"
-	"golang.org/x/net/context"
+	"go.uber.org/zap"
 )
 
-var logrusLogger *logrus.Logger
+var zapLogger *zap.Logger
 
 // Simple unary handler that adds custom fields to the requests's context. These will be used for all log statements.
 func Example_HandlerUsageUnaryPing() {
 	_ = func(ctx context.Context, ping *pb_testproto.PingRequest) (*pb_testproto.PingResponse, error) {
 		// Add fields the ctxtags of the request which will be added to all extracted loggers.
 		grpc_ctxtags.Extract(ctx).Set("custom_tags.string", "something").Set("custom_tags.int", 1337)
-		// Extract a single request-scoped logrus.Logger and log messages.
-		l := ctx_logrus.Extract(ctx)
+
+		// Extract a single request-scoped zap.Logger and log messages.
+		l := ctxzap.Extract(ctx)
 		l.Info("some ping")
 		l.Info("another ping")
 		return &pb_testproto.PingResponse{Value: ping.Value}, nil
