+++ logging/zap/server_interceptors.go
@@ -5,6 +5,7 @@ import (
 	"time"
 
 	"github.com/grpc-ecosystem/go-grpc-middleware"
+	"github.com/grpc-ecosystem/go-grpc-middleware/logging/zap/ctxzap"
 	"github.com/grpc-ecosystem/go-grpc-middleware/tags/zap"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
@@ -85,11 +86,11 @@ func serverCallFields(fullMethodString string) []zapcore.Field {
 }
 
 func newLoggerForCall(ctx context.Context, logger *zap.Logger, fullMethodString string, start time.Time) context.Context {
-	f := ctx_zap.TagsToFields(ctx)
+	f := ctxzap.TagsToFields(ctx)
 	f = append(f, zap.String("grpc.start_time", start.Format(time.RFC3339)))
 	if d, ok := ctx.Deadline(); ok {
 		f = append(f, zap.String("grpc.request.deadline", d.Format(time.RFC3339)))
 	}
 	callLog := logger.With(append(f, serverCallFields(fullMethodString)...)...)
-	return ctx_zap.ToContext(ctx, callLog)
+	return ctxzap.ToContext(ctx, callLog)
 }
