--- tags/logrus/README.md
@@ -1,1 +1,0 @@
-DOC.md
