+++ tags/zap/context.go
@@ -1,64 +1,20 @@
 package ctx_zap
 
 import (
-	"github.com/grpc-ecosystem/go-grpc-middleware/tags"
+	"github.com/grpc-ecosystem/go-grpc-middleware/logging/zap/ctxzap"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 	"golang.org/x/net/context"
 )
 
-type ctxMarker struct{}
-
-type ctxLogger struct {
-	logger *zap.Logger
-	fields []zapcore.Field
-}
-
-var (
-	ctxMarkerKey = &ctxMarker{}
-	nullLogger   = zap.NewNop()
-)
-
 // AddFields adds zap fields to the logger.
+// Deprecated: should use the ctx_zap.AddFields instead
 func AddFields(ctx context.Context, fields ...zapcore.Field) {
-	l, ok := ctx.Value(ctxMarkerKey).(*ctxLogger)
-	if !ok || l == nil {
-		return
-	}
-	l.fields = append(l.fields, fields...)
-
+	ctxzap.AddFields(ctx, fields...)
 }
 
 // Extract takes the call-scoped Logger from grpc_zap middleware.
-//
-// It always returns a Logger that has all the grpc_ctxtags updated.
+// Deprecated: should use the ctx_zap.Extract instead
 func Extract(ctx context.Context) *zap.Logger {
-	l, ok := ctx.Value(ctxMarkerKey).(*ctxLogger)
-	if !ok || l == nil {
-		return nullLogger
-	}
-	// Add grpc_ctxtags tags metadata until now.
-	fields := TagsToFields(ctx)
-	// Add zap fields added until now.
-	fields = append(fields, l.fields...)
-	return l.logger.With(fields...)
-}
-
-// TagsToFields transforms the Tags on the supplied context into zap fields.
-func TagsToFields(ctx context.Context) []zapcore.Field {
-	fields := []zapcore.Field{}
-	tags := grpc_ctxtags.Extract(ctx)
-	for k, v := range tags.Values() {
-		fields = append(fields, zap.Any(k, v))
-	}
-	return fields
-}
-
-// ToContext adds the zap.Logger to the context for extraction later.
-// Returning the new context that has been created.
-func ToContext(ctx context.Context, logger *zap.Logger) context.Context {
-	l := &ctxLogger{
-		logger: logger,
-	}
-	return context.WithValue(ctx, ctxMarkerKey, l)
+	return ctxzap.Extract(ctx)
 }
